Apply copyright notices to all files
Apache 2.0 leaves open the possibility of applying copyrights to
source files, but does not require it. We previously used the NOTICE
file to disambiguate who has copyright, but since we have lots of
different copyright holders, I thought it was best to just apply the
license to each individual file with the correct copyright years.